pyproject.toml
[tool.poetry]
name = "vk-chat-bot"
version = "0.1.0"
description = ""
authors = ["EvilMadSquirrel <minichev.s.l@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
vk-api = "^11.9.9"
python-dotenv = "^0.20.0"
peewee = "^3.15.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
